+++ rust/index/benches/hierarchical_index/mod.rs
@@ -1,4 +1,7 @@
+pub mod common;
+pub mod config;
 pub mod instrumentation;
 pub mod mem_probe;
+pub mod persistance;
 pub mod reader;
 pub mod writer;
