+++ rust/index/benches/hierarchical_index/persistance.rs
@@ -1,0 +1,12 @@
+
+
+// Blockfile prefix constants
+pub const PREFIX_ROOT: &str = "root";
+pub const PREFIX_NEXT_NODE: &str = "next_node";
+pub const PREFIX_DIM: &str = "dim";
+pub const PREFIX_VERSION: &str = "version";
+pub const PREFIX_EMBEDDING: &str = "embedding";
+pub const PREFIX_CENTROID: &str = "centroid";
+
+pub const SINGLETON_KEY: u32 = 0;
+pub const NO_PARENT: u32 = u32::MAX;
